# backend/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ReadinessTaskResponse": {
        "additionalProperties": false,
        "description": "One readiness task, as the founder sees it.",
        "example": {
          "action": "State customer acquisition cost and the period it was measured over, with the spend it was derived from.",
          "audit_run_id": "b8eabf62-1f2e-4c3d-9a8b-7c6d5e4f3a2b",
          "created_at": "2026-08-05T09:14:22Z",
          "dimension": "unit_economics",
          "dimension_score": 41,
          "id": "5f9b1c1e-6f5a-4c3b-9a1f-2d4e6a8c0b12",
          "is_priority": true,
          "requirement": "required",
          "startup_id": "0b6f0f2a-1c2d-4e5f-8a9b-0c1d2e3f4a5b",
          "status": "open",
          "updated_at": "2026-08-05T09:14:22Z"
        },
        "properties": {
          "action": {
            "description": "What to do, in the rubric's own wording, as issued.",
            "title": "Action",
            "type": "string"
          },
          "assessment_attempts": {
            "description": "How many times evidence for this task has been graded.",
            "title": "Assessment Attempts",
            "type": "integer"
          },
          "attempts_remaining": {
            "description": "Graded attempts left before the task locks. At `0` a new upload is refused with `409` and only a SACI admin can reopen it.",
            "title": "Attempts Remaining",
            "type": "integer"
          },
          "audit_run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "The audit run that most recently raised this gap, or `null` if that run has since been deleted. Use it to link the task back to the report that is asking for it.",
            "title": "Audit Run Id"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dimension": {
            "$ref": "#/components/schemas/Dimension",
            "description": "Which rubric dimension raised this gap."
          },
          "dimension_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "description": "The 0-100 score that produced this task, or `null` when the dimension could not be assessed at all. Use it to order and to explain; do not render it to the founder as a grade.",
            "title": "Dimension Score"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_priority": {
            "description": "Whether this is in the short list to show first. An audit marks at most five, one per weak dimension. If nothing is marked, show everything.",
            "title": "Is Priority",
            "type": "boolean"
          },
          "requirement": {
            "$ref": "#/components/schemas/Requirement",
            "description": "`required` blocks investor visibility; `recommended` does not. Derived from the dimension's score against the readiness threshold, not from a model's opinion."
          },
          "startup_id": {
            "format": "uuid",
            "title": "Startup Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/TaskStatus",
            "description": "Where the task sits in the evidence loop. Written by evidence assessment (T3.5), never by a client."
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "startup_id",
          "audit_run_id",
          "dimension",
          "action",
          "requirement",
          "status",
          "dimension_score",
          "is_priority",
          "assessment_attempts",
          "attempts_remaining",
          "created_at",
          "updated_at"
        ],
        "title": "ReadinessTaskResponse",
        "type": "object"
      },
...
      "Requirement": {
        "description": "Whether a founder must close this gap to become investor-visible.\n\nThe distinction is the whole point of the readiness loop: the gate in T3.6\nis \"audit cleared **and** all required tasks passed\", so `RECOMMENDED` is\ngenuinely optional and `REQUIRED` genuinely blocks. Marking everything\nrequired would make the gate unreachable; marking nothing required would\nmake it meaningless.",
        "enum": [
          "required",
          "recommended"
        ],
        "title": "Requirement",
        "type": "string"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "description": "Access token issued by `POST /v1/auth/login`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST API for the FundReady platform. All resources live under `/v1`. Timestamps are ISO 8601 UTC, currencies are ISO 4217 codes, and money is expressed in integer minor units.\n\nEvery error response uses the envelope `{\"error\": {\"code\", \"message\", \"details\"}}`. Branch on `code`, which is stable; never parse `message`.\n\nEvery response carries an `X-Request-ID` header. Quote it in a support report and we can find the exact log line.",
    "summary": "AI business audit and investor-matching platform for SACI Holdings.",
    "title": "FundReady API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/admin/tasks/{task_id}/reopen": {
      "post": {
        "description": "**SACI admins only.** Resets a task's assessment attempts to zero and returns it to `open`, so the founder can submit again.\n\nThis is the only way past the attempt cap, and therefore the only lever that undoes the guard on the investor-visibility gate \u2014 every call is written to the immutable audit log with the admin who made it.\n\nA `passed` task can be reopened too. That is the dispute path for a grading somebody believes was wrong, in either direction.",
        "operationId": "reopen_task_v1_admin_tasks__task_id__reopen_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Reopen a locked task",
        "tags": [
          "readiness"
        ]
      }
    },
...
    "/v1/evidence/{evidence_id}/complete": {
      "post": {
        "description": "Call this after the `PUT` succeeds. The server reads the object back and judges it on what actually arrived, not on what you declared.\n\nOn success the task moves to `submitted` and grading is queued. **Grading is asynchronous** \u2014 poll the task or this submission until `outcome` is non-null. Expect seconds to a minute.\n\nA file that is missing, empty, larger than `max_bytes`, or of an unaccepted type is **deleted from storage** and marked `rejected`; `422` carries the reason in `details.reason` (`object_missing`, `empty_file`, `file_too_large`, `unsupported_content_type`). A rejected upload costs no assessment attempt.\n\nSafe to retry: calling it again on a confirmed submission returns it unchanged and does not queue a second grading.",
        "operationId": "complete_evidence_upload_v1_evidence__evidence_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Evidence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Confirm an evidence upload",
        "tags": [
          "readiness"
        ]
      }
    },
    "/v1/evidence/{evidence_id}/download": {
      "get": {
        "description": "A short-lived signed URL for a file the caller owns, or any file for a SACI admin reviewing a disputed grading.\n\n`404` for a `rejected` submission \u2014 its bytes were deleted at validation \u2014 and for one still `pending`.",
        "operationId": "download_evidence_v1_evidence__evidence_id__download_get",
        "parameters": [
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Evidence Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceDownload"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Download a submission",
        "tags": [
          "readiness"
        ]
      }
    },
...
    "/v1/startups/{startup_id}/tasks": {
      "get": {
        "description": "The tasks this startup's latest audit is asking for, ordered the way to show them: priority items first, then everything that blocks investor visibility, then the rest -- worst-scoring dimension first within each group.\n\n**`requirement` is the field that matters.** `required` blocks investor visibility; `recommended` does not. It is computed from the dimension's score against the readiness threshold, so it can change between audits as a founder improves.\n\n**Empty is a normal answer.** A startup with no succeeded audit has no tasks, and so does one whose audit found no unmet criteria. Neither is an error.\n\n`status` is `open` or `obsolete` today; the evidence states arrive with T3.5. `obsolete` means a later audit stopped raising the gap and nobody had worked on it -- filter it out unless you are showing history.\n\n`total` is the count ignoring pagination, so a client can render \"page N of M\".\n\n**Tasks are generated, never created.** They come from an audit's action plan, so there is no endpoint that creates, edits, or completes one -- a founder becomes investor-visible by doing the work and having evidence assessed (T3.5), not by marking a task done. A task belonging to another founder returns `404`, never `403`: a `403` would confirm the id is real.",
        "operationId": "list_readiness_tasks_v1_startups__startup_id__tasks_get",
        "parameters": [
          {
            "in": "path",
            "name": "startup_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Startup Id",
              "type": "string"
            }
          },
          {
            "description": "Only tasks in this state.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TaskStatus"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only tasks in this state.",
              "title": "Status"
            }
          },
          {
            "description": "Only blocking (`required`) or only optional tasks.",
            "in": "query",
            "name": "requirement",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Requirement"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only blocking (`required`) or only optional tasks.",
              "title": "Requirement"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessTaskPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "List readiness tasks",
        "tags": [
          "readiness"
        ]
      }
    },
    "/v1/startups/{startup_id}/tasks/summary": {
      "get": {
        "description": "How much work stands between this startup and investor visibility, without paging the whole task list -- built for a home screen that loads on every app launch.\n\n`required_open` is the number the visibility gate will test (T3.6): when it reaches zero **and** the audit itself clears, the startup becomes discoverable. Until T3.5 lands, nothing can move a task off `open`, so expect `required_open` to equal `required_total`.\n\nRetired (`obsolete`) tasks are excluded from every count -- they are not work the founder owes.\n\n**Tasks are generated, never created.** They come from an audit's action plan, so there is no endpoint that creates, edits, or completes one -- a founder becomes investor-visible by doing the work and having evidence assessed (T3.5), not by marking a task done. A task belonging to another founder returns `404`, never `403`: a `403` would confirm the id is real.",
        "operationId": "read_readiness_summary_v1_startups__startup_id__tasks_summary_get",
        "parameters": [
          {
            "in": "path",
            "name": "startup_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Startup Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Readiness progress at a glance",
        "tags": [
          "readiness"
        ]
      }
    },
    "/v1/startups/{startup_id}/tasks/{task_id}": {
      "get": {
        "description": "One task in full. `audit_run_id` links back to the report that is asking for it, so a client can show the founder why.\n\n**Tasks are generated, never created.** They come from an audit's action plan, so there is no endpoint that creates, edits, or completes one -- a founder becomes investor-visible by doing the work and having evidence assessed (T3.5), not by marking a task done. A task belonging to another founder returns `404`, never `403`: a `403` would confirm the id is real.",
        "operationId": "read_readiness_task_v1_startups__startup_id__tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "startup_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Startup Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReadinessTaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "not_found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "validation_error"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ],
        "summary": "Read one readiness task",
        "tags": [
          "readiness"
        ]
      }
    },
...
  }
}